config server edit
