+++ examples/config_server/contentconfig/config.json
@@ -1,0 +1,10 @@
+{
+  "analyticsEndpoint": "http://localhost:8080/log",
+  "projectId": "123456789",
+  "contentServingId": "",
+  "contentId": "",
+  "idleTime": 300000,
+  "isPdp": "",
+  "rawContentSelectors": [],
+  "contentSelectors": {}
+}
