+++ examples/config_server/lib/cors.js
