+++ examples/config_server/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "endpoint",
+  "version": "1.0.0",
+  "description": "This server is used to mock API calls for client configuration and optimized content delievery, as well as to send the qeen script.",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "cors": "^2.8.5",
+    "express": "^4.19.2"
+  },
+  "devDependencies": {
+    "@faker-js/faker": "^8.4.1"
+  }
+}
