+++ examples/react/README.md
@@ -3,15 +3,29 @@ This is a demo React application designed to showcase basic functionalities such
 
 ### Backend Server
 Start the backend server:\
+`npm install next`\
 `npm run build`\
 `npm run dev`\
 This will run the server on [http://localhost:3000]
 
+### Config Server
+Start the endpoint server:\
+This endpoint is to mimic our cofing and content endpoint\
+`npm install nodejs`\
+`npm install cors`\
+`node index.js`\
+This will run the server on [http://localhost:4000]
+
 ### Frontend App
 Start the frontend app:\
+`npm install vite`\
 `npm run build`\
 `npm run start`\
 This will run the server on [http://localhost:5000]
+
+### Running All Commands at Once
+In the examples directory, run the following command: \
+`.\run_commands.bat`
 
 ### Usage
 1. Navigate to Home Page: Displays all products.
